Feat: nav바에 결제로직 이동을 위해 portone 로직 전역 상태로 관리
